# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/verify/argo/events": {
            "post": {
                "description": "由 Argo 或外部发布观察器回写 Job 级状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verify"
                ],
                "summary": "回写发布状态",
                "parameters": [
                    {
                        "description": "Release Status Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_api.VerifyReleaseStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/verify/release/steps": {
            "post": {
                "description": "由 Argo Application / Deployment / Rollout 观察器回写 Job steps",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verify"
                ],
                "summary": "回写发布步骤",
                "parameters": [
                    {
                        "description": "Release Step Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_api.VerifyReleaseStepRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/verify/tekton/events": {
            "post": {
                "description": "由 Tekton 或外部构建观察器回写 Manifest 级状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verify"
                ],
                "summary": "回写构建状态",
                "parameters": [
                    {
                        "description": "Build Status Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_api.VerifyBuildStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/verify/tekton/steps": {
            "post": {
                "description": "由 Tekton TaskRun 观察器回写 Manifest steps",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verify"
                ],
                "summary": "回写构建步骤",
                "parameters": [
                    {
                        "description": "Build Step Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_api.VerifyBuildStepRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_bsonger_devflow-app-service_pkg_model.JobStatus": {
            "type": "string",
            "enum": [
                "Pending",
                "Running",
                "Succeeded",
                "Failed",
                "RollingBack",
                "RolledBack",
                "Syncing",
                "SyncFailed"
            ],
            "x-enum-varnames": [
                "JobPending",
                "JobRunning",
                "JobSucceeded",
                "JobFailed",
                "JobRollingBack",
                "JobRolledBack",
                "JobSyncing",
                "JobSyncFailed"
            ]
        },
...
        "github_com_bsonger_devflow-app-service_pkg_model.ManifestStatus": {
            "type": "string",
            "enum": [
                "Pending",
                "Running",
                "Succeeded",
                "Failed"
            ],
            "x-enum-varnames": [
                "ManifestPending",
                "ManifestRunning",
                "ManifestSucceeded",
                "ManifestFailed"
            ]
        },
...
        "github_com_bsonger_devflow-app-service_pkg_model.StepStatus": {
            "type": "string",
            "enum": [
                "Pending",
                "Running",
                "Succeeded",
                "Failed"
            ],
            "x-enum-varnames": [
                "StepPending",
                "StepRunning",
                "StepSucceeded",
                "StepFailed"
            ]
        },
...
        "pkg_api.VerifyBuildStatusRequest": {
            "type": "object",
            "required": [
                "manifest_id",
                "status"
            ],
            "properties": {
                "external_ref": {
                    "type": "string"
                },
                "intent_id": {
                    "type": "string"
                },
                "manifest_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.ManifestStatus"
                }
            }
        },
        "pkg_api.VerifyBuildStepRequest": {
            "type": "object",
            "required": [
                "manifest_id",
                "status",
                "task_name"
            ],
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "manifest_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.StepStatus"
                },
                "task_name": {
                    "type": "string"
                },
                "task_run": {
                    "type": "string"
                }
            }
        },
        "pkg_api.VerifyReleaseStatusRequest": {
            "type": "object",
            "required": [
                "job_id",
                "status"
            ],
            "properties": {
                "external_ref": {
                    "type": "string"
                },
                "intent_id": {
                    "type": "string"
                },
                "job_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.JobStatus"
                }
            }
        },
        "pkg_api.VerifyReleaseStepRequest": {
            "type": "object",
            "required": [
                "job_id",
                "status",
                "step_name"
            ],
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "job_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.StepStatus"
                },
                "step_name": {
                    "type": "string"
                }
            }
        }
    }
}